requirements.txt
opencv-python
scikit-learn
numpy
scikit-image
joblib
Pillow